게시판 boarderAdapter분리
